Changed the type definition of Int_32 in Loris_types.h from long to int, since on alpha a long is 8 bytes, but an int seems to be 4 bytes everywhere. This affects the ability to read AIFF files, and we really need something much better than the Loris_types.h header. We should be able to determine good definitions for these types at config time.

+++ Loris/src/AiffFile.C
@@ -440,7 +440,7 @@ AiffFile::readContainer( std::istream & s )
 
 	//	make sure its really AIFF:
 	if ( ck.formType != AiffType )
-		Throw( FileIOException, std::string("Bad form type in AIFF file: ") + std::string( ck.formType, 4 ) );
+		Throw( FileIOException, std::string("Bad form type in AIFF file: ") + std::string( (long)ck.formType, 4 ) );
 }	
 
 // ---------------------------------------------------------------------------
